+++ src/backend/funmooc/gimporter/import_scripts/blogposts.py
@@ -1,4 +1,7 @@
 # Import blogposts from a Google Sheet
+import re
+from functools import reduce
+
 from django.conf import settings
 from django.utils.text import slugify
 
@@ -6,7 +9,11 @@ from cms.api import create_page
 from cms.models import Page
 from filer.models import Folder
 from richie.apps.courses.cms_plugins import CategoryPlugin
-from richie.apps.courses.defaults import BLOGPOSTS_PAGE
+from richie.apps.courses.defaults import (
+    BLOGPOSTS_PAGE,
+    COURSES_PAGE,
+    ORGANIZATIONS_PAGE,
+)
 from richie.apps.courses.models import BlogPost
 from richie.plugins.plain_text.cms_plugins import PlainTextPlugin
 from richie.plugins.simple_picture.cms_plugins import SimplePicturePlugin
@@ -24,6 +31,34 @@ from .helpers import (
 def import_blogposts(sheet):
     """Import blogposts from a Google Sheet's "blogposts" tab."""
 
+    course_page_path = Page.objects.get(
+        reverse_id=COURSES_PAGE["reverse_id"], publisher_is_draft=True
+    ).get_path()
+
+    # We need to replace old urls by new ones in the blog posts content:
+    # - Get a mapping from organizations old slugs to new paths
+    organizations_mapping = {}
+    for record in sheet.worksheet(ORGANIZATIONS_PAGE["reverse_id"]).get_all_records():
+        try:
+            organizations_mapping[record["slug"].lower()] = Page.objects.get(
+                reverse_id=record["reverse_id"], publisher_is_draft=True
+            ).get_path()
+        except Page.DoesNotExist:
+            pass
+
+    # - Get a mapping from courses old slugs to new paths
+    courses_mapping = reduce(
+        lambda acc, record: {
+            **acc,
+            str(record["reverse_id"]): Page.objects.get(
+                reverse_id=record["reverse_id"], publisher_is_draft=True
+            ).get_path(),
+        },
+        sheet.worksheet(COURSES_PAGE["reverse_id"]).get_all_records(),
+        {},
+    )
+
+    # Get or create the root page for blogposts
     language = settings.LANGUAGE_CODE
     root_reverse_id = BLOGPOSTS_PAGE["reverse_id"]
     root_page = create_page_from_info(root_reverse_id)
@@ -117,7 +152,53 @@ def import_blogposts(sheet):
         # Add a plugin for the body
         placeholder_body = blogpost_page.placeholders.get(slot="body")
         if record["body"]:
-            body = extract_and_replace_media(record["body"])
+            body = extract_and_replace_media(record["body"], folder=blogposts_folder)
+
+            # All urls pointing to this site should be relative
+            # In the content we found a lot of broken urls pointing to:
+            # - fun-mooc.fr
+            # - france-universite-numerique-mooc.fr (already broken on the old site)
+            #
+            # Many derived urls will be found e.g:
+            # - https://fun-mooc.fr            to be replaced by "/"
+            # - http://www.fun-mooc.fr/courses to be replaced by "/courses"
+            # - fun-mooc.fr/courses/my-course  to be replaced by "/courses/my-course"
+            def domain_sub(match):
+                """Return the matched path between quotes if any, or "/" otherwise."""
+                return '"{:s}"'.format(match.group(1)) if match.group(1) else '"/"'
+
+            body = re.sub(
+                (
+                    r"""["|'] ?(?:http)?s?(?:\://)?(?:www\.|studio\.)?"""
+                    r"""(?:france-universite-numerique-mooc|fun-mooc)\.fr([^"']*)["|']"""
+                ),
+                domain_sub,
+                body,
+            )
+
+            # Search and replace universities urls
+            pattern = r'/universities/{:s}/?"'
+            organizations_urls = re.findall(pattern.format(r'([^> "/]*)'), body)
+            for organization_url in organizations_urls:
+                body = re.sub(
+                    pattern.format(organization_url),
+                    f'/{organizations_mapping[organization_url.lower()]:s}/"',
+                    body,
+                )
+
+            # Search and replace courses urls
+            for pattern, group, separator in (
+                (r"/courses/course-v1:{:s}/about", '([^> "/]*)', "+"),
+                (r"/courses/{:s}/about", '([^> "+]*)', "/"),
+            ):
+                courses_urls = re.findall(pattern.format(group), body)
+                for course_url in courses_urls:
+                    course_number = course_url.split(separator)[1].lower()
+                    if course_number in courses_mapping:
+                        replacement = f"/{courses_mapping[course_number]:s}/"
+                    else:
+                        replacement = f"/{course_page_path:s}/"
+                    body = body.replace(pattern.format(course_url), replacement)
 
             create_or_update_single_plugin(
                 placeholder_body, CKEditorPlugin, language=language, body=body